pyproject.toml
[project]
name = "simclr"
version = "0.1.0"
description = "SimCLR: A Simple Framework for Contrastive Learning of Visual Representations"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "torch==2.9.1",
    "torchvision==0.24.1",
    "numpy==2.3.5",
    "tqdm==4.67.1",
    "wandb==0.23.1",
    "scikit-learn==1.8.0",
    "pillow==12.0.0",
    "pyyaml==6.0.3",
    "pydantic==2.12.5",
]

[project.optional-dependencies]
dev = [
    "ruff>=0.1.0",
    "mypy>=1.0.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["sim_clr", "utils"]

[tool.ruff]
line-length = 130
target-version = "py310"

[tool.ruff.lint]
select = ["E", "F", "I", "W"]

[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_configs = true
